tracking: move stories functions from server to lib
